requirements.txt
flask>=2.0.0
flask-cors>=3.0.0
paho-mqtt>=1.6.0

